requirement.txt
pytesseract
pdf2image
Pillow
chromadb
langchain
sentence-transformers